# Conflicts:
#	.config/ags/config.js
#	.config/ags/layouts/topbar.js
#	.config/ags/layouts/unity.js
#	.config/ags/layouts/widgets/desktop.js
#	.config/ags/layouts/widgets/media.js
#	.config/ags/layouts/widgets/quicksettings.js
#	.config/ags/modules/asusctl.js
#	.config/ags/modules/brightness.js
#	.config/ags/scss/variables.scss
#	.config/ags/scss/widgets/bar.scss
#	.config/ags/scss/widgets/desktop.scss
#	.config/ags/scss/widgets/dock.scss
#	.config/ags/scss/widgets/powermenu.scss
#	.config/ags/scss/widgets/settings.scss
#	.config/ags/settings/defaults.js
#	.config/ags/settings/dialog.js
#	.config/ags/settings/service.js
#	.config/nvim/lua/plugins/ui.lua

+++ .config/ags/modules/screenrecord.js
@@ -10,14 +10,17 @@ class RecorderService extends Service {
         Service.register(this, { 'timer': ['int'] });
     }
 
-    _path = GLib.get_home_dir()+'/Videos/Screencasting';
+    _path = GLib.get_home_dir() + '/Videos/Screencasting';
+    _recording = false;
 
     start() {
-        execAsync('slurp', out => {
+        if (this._recording)
+            return;
+
+        execAsync('slurp', area => {
             ensureDirectory(this._path);
             this._file = `${this._path}/${now()}.mp4`;
-            execAsync(['wf-recorder', '-g', out.trim(), '-f', this._file]);
-            print('wf-recorder', '-g', out.trim(), '-f', this._file);
+            execAsync(['wf-recorder', '-g', area.trim(), '-f', this._file]);
             this._recording = true;
             this.emit('changed');
 
@@ -30,6 +33,9 @@ class RecorderService extends Service {
     }
 
     stop() {
+        if (!this._recording)
+            return;
+
         execAsync('killall -INT wf-recorder');
         this._recording = false;
         this.emit('changed');
@@ -43,15 +49,15 @@ class RecorderService extends Service {
             this._file,
         ], res => {
             if (res.trim() === 'files')
-                execAsync('xdg-open '+this._path);
+                execAsync('xdg-open ' + this._path);
 
             if (res.trim() === 'view')
-                execAsync('xdg-open '+this._file);
+                execAsync('xdg-open ' + this._file);
         });
     }
 
     screenshot() {
-        const path = GLib.get_home_dir()+'/Pictures/Screenshots';
+        const path = GLib.get_home_dir() + '/Pictures/Screenshots';
         ensureDirectory(path);
         const file = `${path}/${now()}.png`;
         execAsync(`watershot -c -s path ${file}`);
@@ -68,10 +74,10 @@ class RecorderService extends Service {
                 file,
             ], res => {
                 if (res.trim() === 'files')
-                    execAsync('xdg-open '+path);
+                    execAsync('xdg-open ' + path);
 
                 if (res.trim() === 'view')
-                    execAsync('xdg-open '+file);
+                    execAsync('xdg-open ' + file);
             });
 
             GLib.source_remove(id);
@@ -111,7 +117,7 @@ Widget.widgets['recorder/indicator-button'] = props => Widget({
                 connections: [[Recorder, (label, time) => {
                     const sec = time % 60;
                     const min = Math.floor(time / 60);
-                    label.label = `${min}:${sec < 10 ? '0'+sec : sec}`;
+                    label.label = `${min}:${sec < 10 ? '0' + sec : sec}`;
                 }, 'timer']],
             },
         ],
